Refine selection utilities and consolidate parent generation
