+++ al_pipeline/selection/generate_parents.py
@@ -1,5 +1,4 @@
 import numpy as np
-import matplotlib.pyplot as plt
 import pandas as pd
 import os
 import argparse
